Third refactor of LineItemApp.
Modified 'Product' class: uses float type instead of double type for
'price' variable, getPrice(), and setPrice(). 

Modified 'LineItem' class: getTotal() returns float type instead of
double type. 

Modified 'ProductDB' class: refactored statements using setPrice() to
pass float literals. 

Modified 'LineItemApp' class: added byte variable 'count', added float
variables 'total', 'minItem', and 'maxItem'. Added additonal SYSOUT
statements to print value of number of line items created, invoice
(gross) total, largest line item, and smallest line item to console.

+++ LineItemApp/src/lineItemApp/Product.java
@@ -7,7 +7,7 @@ public class Product {
 	//the instance variables
 	private String code;
 	private String description;
-	private double price;
+	private float price;
 	
 	//the constructor
 	public Product() {
@@ -35,11 +35,11 @@ public class Product {
 	}//close getDescription() method
 	
 	//the set and get methods for the price variable
-	public void setPrice(double price) {
+	public void setPrice(float price) {
 		this.price = price;
 	}//close setPrice() method
 	
-	public double getPrice() {
+	public float getPrice() {
 		return price;
 	}//close getPrice() method
